backend-flask/requirements.txt
flask
flask-cors

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

aws-xray-sdk
watchtower
blinker
rollbar
python-jose